Update stress_test.cpp
- Attempting to make maxhormonediff a double as previously the hormone matrices were integers and the two variable types couldn't interact

+++ stress_test.cpp
@@ -371,7 +371,7 @@ void ReplaceHormone()
         for (s=0;s<maxS;s++)
         {
           tothormonediff = tothormonediff + abs(hormone_next[t][d][s]-hormone[t][d][s]);
-	  maxhormonediff = max(maxhormonediff, abs(hormone_next[t][d][s]-hormone[t][d][s]));
+	  maxhormonediff = max(maxhormonediff, static_cast<double>(abs(hormone_next[t][d][s]-hormone[t][d][s])));
           hormone_next[t][d][s] = hormone[t][d][s];
         }
       }
